rename Toolbar file to BlocksToolbar
