+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/BlocksToolbar.tsx
